[lld][ELF] Add --eh-frame-hdr-format option for sdata8 support
Add a new linker option --eh-frame-hdr-format={sdata4,sdata8} to control
the format used for the .eh_frame_hdr binary search table entries.

The default sdata4 format uses 32-bit signed integers for PC-relative
and FDE VA-relative offsets, which limits the range to +/- 2GiB. For
very large binaries where .text is more than 2GiB from .eh_frame_hdr,
this causes overflow errors.

The new sdata8 format uses 64-bit signed integers, removing this limit.

Changes:
- Add EhFrameHdrFormat enum to Config.h
- Add --eh-frame-hdr-format= option to Options.td
- Parse the option in Driver.cpp
- Update SyntheticSections.cpp to generate sdata8 format when requested
- Update EhFrameHeader::getSize() to return correct size for sdata8
- Update error message to suggest using --eh-frame-hdr-format=sdata8
- Add test for the new option

Note: The sdata8 format requires runtime support (libunwind, libgcc_s)
that can parse the larger entries. This is intended for specialized
use cases with very large binaries.

+++ lld/ELF/SyntheticSections.h
@@ -64,8 +64,8 @@ public:
   size_t numFdes = 0;
 
   struct FdeData {
-    uint32_t pcRel;
-    uint32_t fdeVARel;
+    int64_t pcRel;
+    int64_t fdeVARel;
   };
 
   ArrayRef<CieRecord *> getCieRecords() const { return cieRecords; }
